requirements.txt
APScheduler==3.11.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.2.1
Flask==3.1.1
flask-cors==5.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.1
requests==2.32.5
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.3

psycopg2-binary==2.9.9
gunicorn==21.2.0

